Add uploading portrait image
Debugging is in order as this is the very first iteration
that has functional file input as well as capability
to display the uploaded image in the preview section. No styling
is implemented at all, so the image might be far larger than the preview
section. But hey, it finally works!