Create ContextGrouper using the module pattern
This trick allows to remove all the boilerplate code used for
contexts related tasks from the storage.jsx file. The contexts are now
declared in a single place so all the buttons in the storage.jsx fine need
not those bloody arrays of contexts/dispatches and group names, resulting
in a slightly cleaner codebase. The trick is to only declare the
contexts/dispatches in the ContextGrouper module and call them just in
time, in a map function. This somehow obeys the rule that requires
React hooks to be called on the top level in a component.