Tidy up storage codebase
Split storage.jsx into two files: one for the components (buttons),
the second one for the helper functions that the buttons use to perform
their tasks.

Remove download.jsx. This removes the Download Button (which only alerted()
the user that in order to download the CV, one has to use the browser's
print function and select 'save as pdf' as the output device);

Both changes slightly affect CVForm.jsx file.